moved CubeOps and RollupOps to RelationalGroupsOps, refactored comments
